python_requirements.txt
pyyaml
progressbar2
python-ctags3
pymysql
pyinotify